# openapi.json
{
  "components": {
    "schemas": {
...
      "ChildJobStatus": {
        "additionalProperties": false,
        "description": "Status payload for a child job.",
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "done",
              "error",
              "canceled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status"
        ],
        "title": "ChildJobStatus",
        "type": "object"
      },
...
      "GenerateLessonRequest": {
        "additionalProperties": false,
        "description": "Request payload for lesson generation.",
        "properties": {
          "blueprint": {
            "description": "Required blueprint guidance for lesson planning.",
            "enum": [
              "skillbuilding",
              "knowledgeunderstanding",
              "communicationskills",
              "planningandproductivity",
              "movementandfitness",
              "growthmindset",
              "criticalthinking",
              "creativeskills",
              "webdevandcoding",
              "languagepractice"
            ],
            "title": "Blueprint",
            "type": "string"
          },
          "depth": {
            "default": "highlights",
            "description": "Requested lesson depth (Highlights=2, Detailed=6, Training=10).",
            "enum": [
              "highlights",
              "detailed",
              "training"
            ],
            "title": "Depth",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "maxLength": 300,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional user-supplied details (max 300 characters).",
            "examples": [
              "Focus on lists and loops"
            ],
            "title": "Details"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional client-generated UUID to prevent duplicate processing of the same request.",
            "title": "Idempotency Key"
          },
          "learner_level": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional learner level hint used for prompt guidance.",
            "title": "Learner Level"
          },
          "outcomes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 8,
                "minItems": 1,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional outcomes to guide the planner (3-8 items recommended).",
            "title": "Outcomes"
          },
          "primary_language": {
            "default": "English",
            "description": "Primary language for lesson output.",
            "enum": [
              "English",
              "German",
              "Urdu"
            ],
            "title": "Primary Language",
            "type": "string"
          },
          "schema_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional schema version to pin the lesson output to.",
            "title": "Schema Version"
          },
          "teaching_style": {
            "description": "Required teaching style guidance for lesson planning.",
            "items": {
              "enum": [
                "conceptual",
                "theoretical",
                "practical"
              ],
              "type": "string"
            },
            "maxItems": 3,
            "minItems": 1,
            "title": "Teaching Style",
            "type": "array"
          },
          "topic": {
            "description": "Topic to generate a lesson for.",
            "examples": [
              "Introduction to Python"
            ],
            "minLength": 1,
            "title": "Topic",
            "type": "string"
          },
          "widgets": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 7,
                "minItems": 3,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional list of allowed widgets (overrides defaults).",
            "title": "Widgets"
          }
        },
        "required": [
          "topic",
          "blueprint",
          "teaching_style"
        ],
        "title": "GenerateLessonRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobCreateRequest": {
        "additionalProperties": false,
        "description": "Request payload for creating a background job.",
        "properties": {
          "idempotency_key": {
            "description": "Client-generated key used to deduplicate submissions.",
            "minLength": 1,
            "title": "Idempotency Key",
            "type": "string"
          },
          "job_kind": {
            "enum": [
              "lesson",
              "research",
              "youtube",
              "maintenance",
              "writing",
              "system"
            ],
            "title": "Job Kind",
            "type": "string"
          },
          "lesson_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional lesson identifier bound to this job.",
            "title": "Lesson Id"
          },
          "parent_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional parent job id for child job creation.",
            "title": "Parent Job Id"
          },
          "payload": {
            "additionalProperties": true,
            "description": "Agent-specific payload.",
            "title": "Payload",
            "type": "object"
          },
          "section_id": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional section row id bound to this job.",
            "title": "Section Id"
          },
          "target_agent": {
            "minLength": 1,
            "title": "Target Agent",
            "type": "string"
          }
        },
        "required": [
          "job_kind",
          "target_agent",
          "idempotency_key"
        ],
        "title": "JobCreateRequest",
        "type": "object"
      },
      "JobCreateResponse": {
        "description": "Response payload for job creation.",
        "properties": {
          "expected_sections": {
            "default": 0,
            "description": "Total number of sections expected for lesson jobs.",
            "minimum": 0.0,
            "title": "Expected Sections",
            "type": "integer"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          }
        },
        "required": [
          "job_id"
        ],
        "title": "JobCreateResponse",
        "type": "object"
      },
      "JobRecord": {
        "properties": {
          "artifacts": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifacts"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "completed_section_indexes": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed Section Indexes"
          },
          "completed_sections": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed Sections"
          },
          "completed_steps": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed Steps"
          },
          "cost": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "current_section_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Section Index"
          },
          "current_section_retry_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Section Retry Count"
          },
          "current_section_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Section Status"
          },
          "current_section_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Section Title"
          },
          "expected_sections": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Sections"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "job_kind": {
            "enum": [
              "lesson",
              "research",
              "youtube",
              "maintenance",
              "writing",
              "system"
            ],
            "title": "Job Kind",
            "type": "string"
          },
          "lesson_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lesson Id"
          },
          "logs": {
            "items": {
              "type": "string"
            },
            "title": "Logs",
            "type": "array"
          },
          "max_retries": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Retries"
          },
          "parent_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Job Id"
          },
          "phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phase"
          },
          "progress": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "request": {
            "additionalProperties": true,
            "title": "Request",
            "type": "object"
          },
          "result_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Json"
          },
          "retry_agents": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Agents"
          },
          "retry_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Count"
          },
          "retry_parent_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Parent Job Id"
          },
          "retry_sections": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Sections"
          },
          "section_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Id"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "done",
              "error",
              "canceled"
            ],
            "title": "Status",
            "type": "string"
          },
          "subphase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subphase"
          },
          "target_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Agent"
          },
          "total_steps": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Steps"
          },
          "ttl": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ttl"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "validation": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation"
          }
        },
        "required": [
          "job_id",
          "user_id",
          "job_kind",
          "request",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "JobRecord",
        "type": "object"
      },
      "JobRetryRequest": {
        "additionalProperties": false,
        "description": "Request payload for retrying a failed job.",
        "properties": {
          "agents": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "planner",
                    "section_builder",
                    "repair"
                  ],
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pipeline agents to retry (defaults to full pipeline).",
            "title": "Agents"
          },
          "sections": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "0-based section indexes to retry (defaults to all sections).",
            "title": "Sections"
          }
        },
        "title": "JobRetryRequest",
        "type": "object"
      },
      "JobStatusResponse": {
        "description": "Status payload for a background job.",
        "properties": {
          "child_jobs": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ChildJobStatus"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Child Jobs"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "lesson_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lesson Id"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "done",
              "error",
              "canceled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status"
        ],
        "title": "JobStatusResponse",
        "type": "object"
      },
...
      "PaginatedResponse_JobRecord_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobRecord"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PaginatedResponse[JobRecord]",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/jobs": {
      "get": {
        "description": "List jobs for admins with pagination, filtering, and sorting to control load and exposure.",
        "operationId": "list_jobs_admin_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "queued",
                    "running",
                    "done",
                    "error",
                    "canceled"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "job_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Job Id"
            }
          },
          {
            "in": "query",
            "name": "job_kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Job Kind"
            }
          },
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "in": "query",
            "name": "target_agent",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Agent"
            }
          },
          {
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "default": "created_at",
              "title": "Sort By",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort_order",
            "required": false,
            "schema": {
              "default": "desc",
              "title": "Sort Order",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_JobRecord_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Jobs",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/jobs": {
      "post": {
        "description": "Create a background lesson generation job.",
        "operationId": "create_job_v1_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "description": "Fetch the status and result of a background job.",
        "operationId": "get_job_status_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Job Status",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/cancel": {
      "post": {
        "description": "Request cancellation of a running background job.",
        "operationId": "cancel_job_v1_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/retry": {
      "post": {
        "description": "Retry a failed job with optional section/agent targeting.",
        "operationId": "retry_job_v1_jobs__job_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRetryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Retry Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/v1/lessons/jobs": {
      "post": {
        "description": "Alias route for creating a background lesson generation job.",
        "operationId": "create_lesson_job_v1_lessons_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateLessonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Lesson Job",
        "tags": [
          "lessons"
        ]
      }
    },
...
  }
}